requirements.txt
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
DateTime==4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.4
django-admin-filter==0.6
django-admin-interface==0.16.1
django-advanced-filters==1.3.0
django-allauth==0.44.0
django-braces==1.14.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-colorfield==0.4.1
django-cors-headers==3.6.0
django-crontab==0.7.1
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-money==1.3.1
django-phone-field==1.8.1
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-rest-registration==0.6.0
django-timezone-field==4.1.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
kombu==4.6.11
Markdown==3.3.4
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
py-moneyed==0.8.0
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.4
razorpay==1.2.0
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==1.3.0
whitenoise==5.2.0
zope.interface==5.3.0
